pyproject.toml
[tool.poetry]
name = "first-impression"
version = "0.1.0"
description = ""
authors = ["dcdnc <kuunii06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3"
aiogram-dialog = "^2"
sqlalchemy = { extras = ["asyncio"], version = "^2" }
environs = "^11"
pydantic-settings = "^2"
fluentogram = "^1"
uvloop = "^0.20"
uvicorn = "^0.30"
fastapi = "^0.112"
redis = "^5"
qrcode = "^7"
pyotp = "^2.9"
aiomysql = "^0.2"
cryptography = "^43.0.0"
nats-py = "^2"
orjson = "^3.10.6"

[tool.poetry.group.dev.dependencies]
locust = "^2"
i18n = "^0.2"
alembic = "^1"
pytest = "^8"
mypy = "^1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs/bot"
no_warn_no_return = true
exclude = ["tests"]

[mypy-bot]
ignore_missing_imports = true
